creates desktop nav component, adds dialog description element
